+++ scanpointgenerator/generators/arraygenerator.py
@@ -13,31 +13,33 @@
 #
 ###
 
+from annotypes import Anno, Union, Array, Sequence
+
 from scanpointgenerator.compat import np
-from scanpointgenerator.core import Generator
+from scanpointgenerator.core import Generator, UAxes, UUnits, AAlternate
+
+with Anno("The array positions"):
+    APoints = Array[np.float64]
+UPoints = Union[APoints, Sequence[np.float64], Sequence[float]]
 
 
-@Generator.register_subclass("scanpointgenerator:generator/ArrayGenerator:1.0")
+@Generator.register_subclass(
+    "scanpointgenerator:generator/ArrayGenerator:1.0")
 class ArrayGenerator(Generator):
-    """Generate points fron a given list of positions"""
+    """Generate points from a given list of positions"""
 
-    def __init__(self, axis, units, points, alternate=False):
-        """
-        Args:
-            axis (str): The scannable axis name
-            units (str): The scannable units.
-            points (list(double)): array positions
-            alternate (bool): Alternate directions
-        """
-
-        self.alternate = alternate
-        self.points = np.array(points, dtype=np.float64)
-        self.size = len(self.points)
-        self.axes = [axis]
-        self.units = {axis:units}
+    def __init__(self, axes, units, points, alternate=False):
+        # type: (UAxes, UUnits, UPoints, AAlternate) -> None
+        super(ArrayGenerator, self).__init__(
+            axes, units, len(points), alternate)
+        assert len(self.axes) == len(self.units) == 1, \
+            "Expected 1D, got axes %s and units %s" % (list(self.axes),
+                                                       list(self.axes))
+        self.points = APoints(np.array(points, dtype=np.float64))
 
     def prepare_arrays(self, index_array):
-        points = self.points
+        # Get the actual numpy array from the Array class wrapper
+        points = self.points.seq
         # add linear extension to ends of points, representing t=-1 and t=N+1
         v_left = points[0] - (points[1] - points[0])
         v_right = points[-1] + (points[-1] - points[-2])
@@ -51,35 +53,6 @@ class ArrayGenerator(Generator):
         index_floor = np.floor(index_array).astype(np.int32)
         epsilon = index_array - index_floor
         index_floor += 1
-        values = points[index_floor] + epsilon * (points[index_floor+1] - points[index_floor])
-        return {self.axes[0]:values}
-
-    def to_dict(self):
-        """Serialize ArrayGenerator to dictionary"""
-
-        d = {
-                "typeid":self.typeid,
-                "axis":self.axes[0],
-                "units":self.units[self.axes[0]],
-                "points":self.points.tolist(),
-                "alternate":self.alternate,
-            }
-        return d
-
-    @classmethod
-    def from_dict(cls, d):
-        """
-        Create a ArrayGenerator from serialized form.
-
-        Args:
-            d (dict): Serialized generator
-
-        Returns:
-            ArrayGenerator: New ArrayGenerator instance
-        """
-
-        axis = d["axis"]
-        units = d["units"]
-        alternate = d["alternate"]
-        points = d["points"]
-        return cls(axis, units, points, alternate)
+        values = points[index_floor] + epsilon * \
+                 (points[index_floor+1] - points[index_floor])
+        return {self.axes[0]: values}