+++ scanpointgenerator/rois/linear_roi.py
@@ -11,22 +11,33 @@
 #
 ###
 
+from annotypes import Anno, Union, Array, Sequence
+
 from math import cos, sin
 
 from scanpointgenerator.core import ROI
 from scanpointgenerator.compat import np
+
+with Anno("The start of the line"):
+    AStart = Array[float]
+UStart = Union[AStart, Sequence[float]]
+with Anno("The length of the line"):
+    ALength = float
+with Anno("The angle of the line"):
+    AAngle = float
 
 
 @ROI.register_subclass("scanpointgenerator:roi/LinearROI:1.0")
 class LinearROI(ROI):
 
     def __init__(self, start, length, angle):
+        # type: (UStart, ALength, AAngle) -> None
         super(LinearROI, self).__init__()
         if length == 0:
             raise ValueError("Line must have non-zero length")
-        self.start = start
-        self.length = length
-        self.angle = angle
+        self.start = AStart(start)
+        self.length = ALength(length)
+        self.angle = AAngle(angle)
 
     def contains_point(self, point, epsilon=1e-15):
         # line's vector is (cphi, sphi)
@@ -81,15 +92,3 @@ class LinearROI(ROI):
         y *= y
         mask |= x + y <= epsilon * epsilon
         return mask
-
-    def to_dict(self):
-        d = super(LinearROI, self).to_dict()
-        d["start"] = self.start
-        d["length"] = self.length
-        d["angle"] = self.angle
-        return d
-
-    @classmethod
-    def from_dict(cls, d):
-        return cls(d["start"], d["length"], d["angle"])
-